I assume we will never have more than 256 players per team

+++ crates/bevyhavior_simulator/src/aufstellung.rs
@@ -6,9 +6,9 @@ use spl_network_messages::{bindings::MAX_NUM_PLAYERS, Penalty, Team};
 use crate::{game_controller::GameControllerCommand, robot::Robot};
 
 pub fn hulks_aufstellung(
-    active_field_players: Vec<usize>,
-    picked_up_players: Vec<usize>,
-    goal_keeper_jersey_number: usize,
+    active_field_players: Vec<u8>,
+    picked_up_players: Vec<u8>,
+    goal_keeper_jersey_number: u8,
     mut commands: Commands,
     game_controller_commands: &mut EventWriter<GameControllerCommand>,
 ) {
@@ -27,7 +27,7 @@ pub fn hulks_aufstellung(
             },
         ));
     }
-    for jersey_number in 1..=MAX_NUM_PLAYERS as usize {
+    for jersey_number in 1..=MAX_NUM_PLAYERS {
         if !active_players_in_game_controller.contains(&jersey_number) {
             game_controller_commands.send(GameControllerCommand::Penalize(
                 jersey_number,
