+++ crates/types/src/filtered_game_controller_state.rs
@@ -15,14 +15,14 @@ pub struct FilteredGameControllerState {
     pub kicking_team: Team,
     pub penalties: Players<Option<Penalty>>,
     pub opponent_penalties: Players<Option<Penalty>>,
-    pub goal_keeper_number: usize,
-    pub opponent_goal_keeper_number: usize,
+    pub goal_keeper_number: u8,
+    pub opponent_goal_keeper_number: u8,
     pub remaining_number_of_messages: u16,
     pub sub_state: Option<SubState>,
     pub own_team_is_home_after_coin_toss: bool,
 
-    pub new_own_penalties_last_cycle: HashMap<usize, Penalty>,
-    pub new_opponent_penalties_last_cycle: HashMap<usize, Penalty>,
+    pub new_own_penalties_last_cycle: HashMap<u8, Penalty>,
+    pub new_opponent_penalties_last_cycle: HashMap<u8, Penalty>,
 }
 
 impl Default for FilteredGameControllerState {
